+++ popcorn/__main__.py
@@ -7,16 +7,16 @@ Args:
 """
 
 import argparse
+import os
 from openpyxl import Workbook
 import sys
 
-from popcorn.analyzers import hotspots, kernel_differences
 from popcorn.interfaces import Verbosity, Kettle, MDTables, CSVArchive
 from popcorn.reporters import report_hotspots, report_kdiff
 from popcorn.readers import LevelZeroTracerJsonReader
 from popcorn.structures import Case
 
-__version__ = "0.0.1"
+__version__ = "0.0.2"
 
 
 def main_cli() -> str | None:
@@ -31,6 +31,12 @@ def main_cli() -> str | None:
         nargs="+",
         type=str,
         help="input trace files (provide at least one)",
+    )
+
+    parser.add_argument("-f", "--folder", "--dir", "--directory",
+        action="store_true",
+        help="input name(s) reference(s) folder(s) and not file(s)",
+        dest="folder_input",
     )
 
     parser.add_argument(
@@ -95,9 +101,33 @@ def main_cli() -> str | None:
 
     args = parser.parse_args()
 
-    input_file_count = len(
-        args.files
-    )  # argparse will ensure at least one file is given
+    reader = (
+        LevelZeroTracerJsonReader()
+    )  # TODO: add more input file formats? and add 'input_type' option to control manually? and autodetect?
+
+    if args.folder_input:
+        args.folders: list[str] = args.files
+        args.files: list[str] = []
+        for folder in args.folders:
+            if os.path.exists(os.path.abspath(folder)):
+                supported_files: list[str] = []
+                for (dirpath, _, filenames) in os.walk(folder):
+                    supported_files.extend([
+                        os.path.join(dirpath, filename) for filename in 
+                        filter(lambda f: f.endswith(reader.format), filenames)
+                    ])
+                if len(supported_files) < 1 and len(args.folders) > 1:
+                    print(f"Could not find any supported files within {folder}!\n")
+                else:
+                    args.files.extend(supported_files)
+            else:
+                print(f"{folder} does not exist!\n")
+    
+    input_file_count = len(args.files)
+    # argparse will ensure args.files > 0 when input type isn't a folder
+    if input_file_count < 1:
+        # input was a foldername which had no supported files in it
+        return f"Error! No supported file(s) found in {' or '.join(args.folders)}!\n"
 
     output_verbosity = Verbosity.STANDARD
     if args.verbose:
@@ -132,10 +162,7 @@ def main_cli() -> str | None:
         )
 
     # extract cases from json files
-    cases = []
-    reader = (
-        LevelZeroTracerJsonReader()
-    )  # TODO: add more input file formats? and add 'input_type' option to control manually? and autodetect?
+    cases: list[Case] = []
     for input_filename in args.files:
         cases.append(
             Case(
@@ -164,17 +191,17 @@ def main_cli() -> str | None:
 
     match args.a:
         case "all":
-            report_hotspots(hotspots(cases), report)
-            report_kdiff(kernel_differences(cases), report)
+            report_hotspots(cases, report)
+            report_kdiff(cases, report)
         case "kdiff":
-            report_kdiff(kernel_differences(cases), report)
+            report_kdiff(cases, report)
         case "pops+kdiff":
-            report_hotspots(hotspots(cases), report)
-            report_kdiff(kernel_differences(cases), report)
+            report_hotspots(cases, report)
+            report_kdiff(cases, report)
         case "pops":
-            report_hotspots(hotspots(cases), report)
+            report_hotspots(cases, report)
         case _:  # default is just hotspots
-            report_hotspots(hotspots(cases), report)
+            report_hotspots(cases, report)
 
     report.save(args.o)
     return None